Corrections on CreateQVD.md
I tried the codes and didn't get the expected result on QlikView (even with the change on lib connection that we don't have in View). 
It can be fixed with just small changes like removing the lonely ' " ' before 'Generating... on TRACE command.
And not sure if is needed but I also removed the ';' on 'Sub' and 'EndSub' lines commands like examples in Qlik Help.